health-data-backend/requirements.txt
anyio==4.10.0
base58==2.1.1
blinker==1.9.0
certifi==2025.8.3
cffi==1.17.1
click==8.2.1
cryptography==45.0.6
Deprecated==1.2.18
ECPy==1.2.5
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pycparser==2.22
pycryptodome==3.23.0
sniffio==1.3.1
SQLAlchemy==2.0.41
types-Deprecated==1.2.15.20250304
typing_extensions==4.14.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
xrpl-py==4.3.0
